+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feat(nuxt): Add `import-in-the-middle` install step when using pnpm ([#727](https://github.com/getsentry/sentry-wizard/pull/727))
 - fix(nuxt): Remove unused parameter in sentry-example-api template ([#734](https://github.com/getsentry/sentry-wizard/pull/734))
+- fix(nuxt): Remove option to downgrade override nitropack ([#744](https://github.com/getsentry/sentry-wizard/pull/744))
 - feat(flutter): Add Flutter support ([#735](https://github.com/getsentry/sentry-wizard/pull/735))
 
 ## 3.36.0